Add Profile template argument to dispatcher.
The profile template argument can be used to prevent generation of
profiling code for a dispatch call, which is a good idea when your
dispatch handler is very small and you shouldn't spend time profiling
it.

While doing this, I also restructured redispatch so it also was
a method on TypedOperatorHandle, and rewrote some call sites to use
the direct method, which shortened some of our generated code.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 98ae4d7261a2487a8a34d5d24e96cd0c3105a7cd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44053