DeviceGuard dispatch key
It turned out this was pretty easy to do after the codegen rewrite, so
I wanted to give it a try in case there were some performance benefits.

The basic concept is to remove DeviceGuard from the CPUType/CUDAType/TypeDefault
codegen (taking us one step closer) and give it its own dispatch key and
codegen.  Unlike the old implementation, where guarding only occurred
immediately before we actually did a backend implementation, in the new
implementation guarding happens as early as possible, and once we've done it
once, we exclude DeviceGuard from the dispatch key set, so that we never
(redundantly) DeviceGuard again.

Need to see if there is a performance benefit or not.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: efb83f47bc43870b52a9a4c4342f483144799b23
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44033