Add ivalue conversion helpers to rpc::Message
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47192

To make progress on retiring torch/csrc/utils/future.h as mentioned in
https://github.com/pytorch/pytorch/issues/41574, we add some helpers to convert
back and forth from message to ivalue tuple representation of the message. We
can then use these APIs in order to convert messages to ivalue, and then wrap
then with at::IValue::Future instead of using torch::utils::Future.

In particular, we introduce `toIValueTuple` which converts the message to an
ivalue tuple and a static `Message::fromIValueTuple` which does the opposite.

Tested by adding a test to test_wire_serialization.cpp
ghstack-source-id: 115776116

Differential Revision: [D24682943](https://our.internmc.facebook.com/intern/diff/D24682943/)