[NNAPI] Handle binary ops combining NHWC+NCHW in some cases
Summary:
This came up in a squeeze-and-excitation model.  Starting with an NHWC
tensor T, we perform a mean operation across H and W, giving an NxC
tensor, which (after some fully connected layers) is reshaped to
NxCx1x1, then multiplied with T.  To handle this, we detect the specific
case of a binary op with one NHWC input and one contiguous input with
H,W == 1,1 and allow the op to be applied (after transposing the
contiguous input).

Test Plan:
Unit test.

ghstack-source-id: 3dd06e312f55c9fccb6537522b39ecf973174c06
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48812