fx quant: add an example for reference quantized bmm
Summary:

Currently, we do not have a quantized bmm kernel. However,
there is a need for people to emulate the numerics of quantized
bmm.

This PR adds an example flow of emulating quantized bmm
with dq -> bmm_fp32 -> q.  We define the interface of the edges
as int8, and use FX graph mode's default observer insertion
logic to insert observers in the right places.  We then
create a custom handler which swaps in a custom module
which emulates the quantized bmm.

There are plans to provide a more generic version version
of this functionality at the framework level in the future,
but custom handlers should solve the problem in the short term.

Test Plan:

```
python test/test_quantization.py TestQuantizeFx.test_reference_bmm_subgraph
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 361b8e1e9cca30225566cb84585fa8ecbb750912
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50313