[not for land] temp fix for per-channel observers used by multiple layers
Summary:

If a layer is used twice, the scale+zp from the first use need to be
passed to the fq backward.  This PR unbreaks this, for now, by cloning
scale + zp.

This is not for land, a better approach would be to stop
passing scale+zp to the backward altogether, we'll put up a separate PR
for that, but it will take more time.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 5404269b9d327b596f9379722844422e0967abf5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50828