Skip resize_output call when TensorIterator has proven it's unnecessary
Previously, we unconditionally called resize_output, even when it was
known (via the op.will_resize boolean) that you have exactly the right
size already.  Thanks to Natalia Gimelshein for diagnosing the problem
and Scott Wolchok for convincing me to fix it this way (as opposed
to speeding up resize_output).

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 5f9089d6efbe160e0d082fb2eaaa8ff188d09395
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50958