Let child CUDAFuture wait for parent CUDAFuture's CUDAEvents
If the child cannot extract tensors from returned IValue, the
current child CUDAFuture won't wait for anything. In this case,
if the `wait()` wasn't called on the parent Future, streams are
not synchronized, and it is possible that parent Future's CUDA
ops have not been added to streams yet.

This commit let the child CUDAFuture holds on to parent's
CUDAEvents at initialization time, and only replaces those events
with its own ones if sub-value extraction from the returned IValue
succeeded.

ghstack-source-id: 8a59f6c25c18c05d93c8a7181032f66e758a7cf4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51506