[JIT] Add `torch.jit.list`
**Summary**
This commit adds `torch.jit.list`, a function that makes lists that have reference
semantics across the Python/TorchScript boundary. That is, modifications
made in TorchScript to instances of lists created using `torch.jit.list`
are visible in Python even when it is not returned from the function.

`torch.jit.list` returns objects of type `torch._C.ScriptList`, which is
implemented using a modified version of pybind's `stl_bind.h`-style bindings
attached to `ScriptList` and `ScriptListIterator`, wrapper classes around
`c10::impl::GenericList` and `c10::impl::GenericList::iterator`. These
bindings allow instances of `torch._C.ScriptList` to be used as if it were a
regular `list` in Python. Reference semantics are achieved by simply
retrieving the `IValue` contained in `ScriptList` in `toIValue` (invoked
when converting Python arguments to `IValues` before calling TorchScript
code).

**Test Plan**
This commit adds `TestScriptList` to `test_list_dict.py`, a set of tests
that check that all of the common list operations are supported
and that instances have reference semantics across the
Python/TorchScript boundary.

ghstack-source-id: 7d1db721430d2f17b96799dcc0131e7211ab1c77
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52832